refactor: extract code to viewport, game, title, and text files
