+++ internal/game/ebiten_renderer.go
@@ -11,25 +11,7 @@ import (
 	"github.com/hajimehoshi/ebiten/v2/text/v2"
 )
 
-const (
-	tileSize               = 16
-	mapViewportWidth       = 56
-	mapViewportHeight      = 20
-	hudPanelWidth          = 24
-	messageLogHeight       = 4
-	titleScreenSubtitle    = "A Cyberpunk Roguelike"
-	titleScreenCopyright   = "Copyright 2025"
-	titleScreenInstruction = "Press SPACE to start or Q to quit"
-)
-
-var titleScreenArt = []string{
-	"  _________                          .__                                          ",
-	" /   _____/_________________ __  _  _|  |_______ __ __  ____   ____   ___________ ",
-	" \\_____  \\\\____ \\_  __ \\__  \\\\ \\/ \\/ /  |\\_  __ \\  |  \\/    \\ /    \\_/ __ \\_  __ \\",
-	" /        \\  |_> >  | \\// __ \\\\     /|  |_|  | \\/  |  /   |  \\   |  \\  ___/|  | \\/",
-	"/_______  /   __/|__|  (____  /\\/\\_/ |____/__|  |____/|___|  /___|  /\\___  >__|   ",
-	"        \\/|__|              \\/                             \\/     \\/     \\/       ",
-}
+const tileSize = 16
 
 // EbitenRenderer handles rendering a Game using the Ebiten game engine.
 type EbitenRenderer struct {
@@ -172,131 +154,4 @@ func (renderer *EbitenRenderer) Draw(screen *ebiten.Image) {
 // Layout returns the game's logical screen size. Required by ebiten.Game interface.
 func (renderer *EbitenRenderer) Layout(outsideWidth, outsideHeight int) (int, int) {
 	return renderer.screenWidth, renderer.screenHeight
-}
-
-// renderGlyph draws a single character glyph at the specified position with the given color.
-// This is a helper method used by RenderTile and RenderPlayer.
-func (renderer *EbitenRenderer) renderGlyph(screen *ebiten.Image, glyph rune, tileX, tileY int, color color.Color) {
-	// Convert tile coordinates to pixel coordinates
-	pixelX := float64(tileX * renderer.tileSize)
-	pixelY := float64(tileY * renderer.tileSize)
-
-	// Draw the glyph
-	glyphString := string(glyph)
-	options := &text.DrawOptions{}
-	options.GeoM.Translate(pixelX, pixelY)
-	options.ColorScale.ScaleWithColor(color)
-
-	text.Draw(screen, glyphString, renderer.fontFace, options)
-}
-
-// RenderTile draws a single tile glyph at the specified tile coordinates.
-// tileX and tileY are in tile units which are converted to pixel coordinates.
-func (renderer *EbitenRenderer) RenderTile(screen *ebiten.Image, tile Tile, tileX, tileY int) {
-	renderer.renderGlyph(screen, tile.Glyph, tileX, tileY, tile.Color)
-}
-
-// RenderPlayer draws the player character at their viewport relative position.
-func (renderer *EbitenRenderer) RenderPlayer(screen *ebiten.Image, player Player) {
-	screenX, screenY := renderer.CalculatePlayerScreenPosition()
-	renderer.renderGlyph(screen, player.Glyph, screenX, screenY, player.Color)
-}
-
-// RenderMap draws all the tiles from the game map that are visible in the viewport.
-func (renderer *EbitenRenderer) RenderMap(screen *ebiten.Image, game *Game) {
-	minX, minY, maxX, maxY := renderer.CalculateViewportBounds()
-
-	for y := minY; y < maxY; y++ {
-		for x := minX; x < maxX; x++ {
-			tile := game.Tiles[y][x]
-
-			// Render at screen position offset by viewport origin
-			screenX := x - minX
-			screenY := y - minY
-
-			renderer.RenderTile(screen, tile, screenX, screenY)
-		}
-	}
-}
-
-// RenderStatsPanel draws the player stats in the right panel (24 columns).
-func (renderer *EbitenRenderer) RenderStatsPanel(screen *ebiten.Image) {
-	// Panel starts at x=56 (after viewport), top of screen
-	panelX := float64(mapViewportWidth * renderer.tileSize)
-	startY := 0.0
-	lineHeight := float64(renderer.tileSize)
-
-	// Draw panel title
-	renderer.drawText(screen, "== Runner ==", panelX, startY, colorYellow)
-
-	// Draw player name
-	nameY := startY + lineHeight*2
-	renderer.drawText(screen, renderer.game.Player.Name, panelX, nameY, colorWhite)
-
-	// Draw level
-	levelY := nameY + lineHeight*2
-	levelText := fmt.Sprintf("Level: %d", renderer.game.Player.Level)
-	renderer.drawText(screen, levelText, panelX, levelY, colorWhite)
-
-	// Draw health
-	healthY := levelY + lineHeight
-	healthText := fmt.Sprintf("Health: %d", renderer.game.Player.Health)
-	renderer.drawText(screen, healthText, panelX, healthY, color.White)
-}
-
-// RenderMessageLog draws the message log area at the bottom of the screen
-// (4 lines high).
-func (renderer *EbitenRenderer) RenderMessageLog(screen *ebiten.Image) {
-	// Message log starts below the viewport (20 tiles down)
-	logY := mapViewportHeight
-
-	// Draw separator line
-	for x := 0; x < renderer.game.Width; x++ {
-		renderer.renderGlyph(screen, '=', x, logY, colorYellow)
-	}
-
-	// If quit confirmation is active show it in the message log
-	if renderer.game.IsConfirmingQuit() {
-		promptX := 1.0 * float64(renderer.tileSize)
-		promptY := float64((logY + 1) * renderer.tileSize)
-		renderer.drawText(screen, "Really quit? (Y/N)", promptX, promptY, colorYellow)
-	}
-}
-
-// drawText is a helper to render text at pixel coordinates.
-func (renderer *EbitenRenderer) drawText(screen *ebiten.Image, txt string, x, y float64, clr color.Color) {
-	options := &text.DrawOptions{}
-	options.GeoM.Translate(x, y)
-	options.ColorScale.ScaleWithColor(clr)
-
-	text.Draw(screen, txt, renderer.fontFace, options)
-}
-
-// RenderTitleScreen draws the title screen with ASCII art and instructions.
-func (renderer *EbitenRenderer) RenderTitleScreen(screen *ebiten.Image) {
-	screen.Fill(colorBlack) // Clear screen to black
-
-	lineHeight := float64(renderer.tileSize)
-	startY := 5.0 * lineHeight
-
-	// Draw title ASCII art
-	for i, line := range titleScreenArt {
-		y := startY + float64(i)*lineHeight
-		renderer.centerText(screen, line, y, colorYellow)
-	}
-
-	metaY := startY + float64(len(titleScreenArt)+3)*lineHeight
-	renderer.centerText(screen, titleScreenSubtitle, metaY, colorWhite)
-	renderer.centerText(screen, titleScreenCopyright, metaY+lineHeight*2, colorWhite)
-
-	instructionsY := float64(renderer.game.Height-3) * lineHeight
-	renderer.centerText(screen, titleScreenInstruction, instructionsY, colorYellow)
-}
-
-// centerText measures, centers, and draws at the given Y position.
-func (renderer *EbitenRenderer) centerText(screen *ebiten.Image, txt string, y float64, clr color.Color) {
-	screenWidthPixels := float64(renderer.game.Width * renderer.tileSize)
-	textWidth := text.Advance(txt, renderer.fontFace)
-	x := (screenWidthPixels - textWidth) / 2.0
-	renderer.drawText(screen, txt, x, y, clr)
 }
